==> notebook/01_extract_osm_libraries.ipynb <==
# Extract Berlin Libraries from OpenStreetMap (Overpass API)

This notebook extracts all libraries in Berlin from OpenStreetMap using **OSMnx**.

- Reasearch, discover, collecte, and evaluating datasets
- Filter tag: `amenity=library`
- Output: Raw GeoDataFrame with OSM attributes (in both .csv and .geojson format)
- CRS: EPSG:4326


# imports

import osmnx as ox                  # fetch data from OpenStreetMap
import geopandas as gpd             # handle geospatial (map) data


# Function Definition
def extract_osm_libraries(place: str = "Berlin, Germany") -> gpd.GeoDataFrame: # Define the place for libraries
    """
    Fetch libraries from OSM for a given place.
    """
    tags = {"amenity": "library"}   # Define the tag for libraries
    gdf = ox.features_from_place(place, tags) # Fetch the data as a GeoDataFrame
    gdf = gdf.to_crs(epsg=4326)   # Convert to WGS84 coordinate system/CRS (that is EPSG:4326)
    gdf = gdf.reset_index()  # Reset index to get the OSM IDs as a column for easier handling
    return gdf


# Execute Extraction
libraries_gdf = extract_osm_libraries()
libraries_gdf.head()


libraries_gdf.describe(include='all').T

# List the columns names of the GeoData

libraries_gdf.columns.to_list()

# Save Output as GeoJSON (with geometry)
output_path = "../sources/osm_libraries.geojson"   #  (includes geometry)
libraries_gdf.to_file(output_path, driver="GeoJSON")

# Save Output as CSV (tabular only, no geometry)
csv_output_path = "../sources/osm_libraries.csv"
libraries_gdf.to_csv(csv_output_path, index=False)

print(f"Extracted {len(libraries_gdf)} libraries")
